+++ sjf4j-testbench/src/test/java/org/sjf4j/testbench/node/ReflectUtilTest.java
@@ -188,7 +188,7 @@ class ReflectUtilTest {
         ) {}
     }
 
-    // ---------- 6) lambdaSetter：reference/primitive/private ----------
+    // ---------- 6) lambdaSetter: reference/primitive/private ----------
 
     @Test
     void lambdaSetter_setsPrivateReferenceField() throws Throwable {