+++ sjf4j/src/test/java/org/sjf4j/node/ReflectUtilTest.java
@@ -161,7 +161,7 @@ class ReflectUtilTest {
         ) {}
     }
 
-    // ---------- 6) lambdaSetter：reference/primitive/private ----------
+    // ---------- 6) lambdaSetter: reference/primitive/private ----------
 
     @Test
     void lambdaSetter_setsPrivateReferenceField() throws Throwable {